rename: 기존 컴포넌트 이름 변경
